+++ ghostel-eshell.el
@@ -87,7 +87,7 @@ eshell."
         (setq-local ghostel-kill-buffer-on-exit
                     (bound-and-true-p eshell-destroy-buffer-when-process-dies))
         (unless ghostel-eshell-track-title
-          (setq-local ghostel-enable-title-tracking nil))
+          (setq-local ghostel-set-title-function nil))
         (add-hook 'ghostel-exit-functions
                   #'ghostel-eshell--visual-exit nil t))
       nil)))